[btree] setup playground for inoder traversal
