+++ explore/binary-tree.playground/Pages/1.Binary Tree Preorder Traversal .xcplaygroundpage/Contents.swift
@@ -6,22 +6,6 @@ import Foundation
 
 */
 
-public class TreeNode : CustomStringConvertible {
-  public var val: Int
-  public var left: TreeNode?
-  public var right: TreeNode?
-
-  public init(_ val: Int, left: TreeNode? = nil, right: TreeNode? = nil) {
-    self.val = val
-    self.left = left
-    self.right = right
-  }
-  
-  public var description: String {
-    return "\(val) -> left: \(String(describing: left)), right: \(String(describing: right))"
-  }
-}
-
 class Solution {
     func preorderTraversal(_ root: TreeNode?) -> [Int] {
       guard let root = root else {
