+++ explore/binary-tree.playground/Pages/2.Binary Tree Inorder Traversal .xcplaygroundpage/Contents.swift
@@ -1,0 +1,25 @@
+//: [Previous](@previous)
+
+import Foundation
+
+/*:
+ ## Binary Tree Inorder Traversal
+ <https://leetcode.com/explore/learn/card/data-structure-tree/134/traverse-a-tree/929/>
+ */
+
+class Solution {
+  func inorderTraversal(_ root: TreeNode?) -> [Int] {
+    var ret: [Int] = []
+    
+    return ret
+  }
+}
+
+let root = TreeNode(1, left: nil,
+                      right: TreeNode(2, left: TreeNode(3, left: nil, right: nil), right: nil
+  )
+)
+
+Solution().inorderTraversal(root)
+
+//: [Next](@next)
