--- explore/binary-tree.playground/Pages/Untitled Page 2.xcplaygroundpage/Contents.swift
@@ -1,9 +1,0 @@
-//: [Previous](@previous)
-
-import Foundation
-
-
-
-
-
-//: [Next](@next)
