+++ explore/binary-tree.playground/contents.xcplayground
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <playground version='6.0' target-platform='ios'>
     <pages>
-        <page name='Untitled Page 2'/>
+        <page name='2.Binary Tree Inorder Traversal '/>
         <page name='1.Binary Tree Preorder Traversal '/>
     </pages>
 </playground>
